- поправил обработку ошибок

+++ controllers/cards.js
@@ -1,3 +1,4 @@
+const { isValidObjectId } = require('mongoose');
 const Card = require('../models/card');
 
 const createCard = (req, res) => {
@@ -22,8 +23,20 @@ const getCards = (req, res) => {
 };
 
 const deleteCard = (req, res) => {
+  if (!isValidObjectId(req.params.cardId.trim().toString())) {
+    res.status(400).send({ message: 'Переданы некорректные данные' });
+    return;
+  }
+
   Card.findByIdAndRemove(req.params.cardId)
-    .then((data) => res.send({ data }))
+    .then((card) => {
+      if (!card) {
+        res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
+        return;
+      }
+
+      res.send({ card });
+    })
     .catch((err) => {
       if (err.name === 'CastError') {
         res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
@@ -35,6 +48,11 @@ const deleteCard = (req, res) => {
 };
 
 const setCardLike = (req, res) => {
+  if (!isValidObjectId(req.params.cardId.trim().toString())) {
+    res.status(400).send({ message: 'Переданы некорректные данные' });
+    return;
+  }
+
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -64,6 +82,11 @@ const setCardLike = (req, res) => {
 };
 
 const deleteCardLike = (req, res) => {
+  if (!isValidObjectId(req.params.cardId.trim().toString())) {
+    res.status(400).send({ message: 'Переданы некорректные данные' });
+    return;
+  }
+
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
